ElixirDropbox v0.0.8 ElixirDropbox.FileRequests

Functions

create(client, title, destination, deadline, open \\ true)

Creates a file request for this user.

Example

deadline = %{ “deadline” => “2020-10-12T17:00:00Z” } ElixirDropbox.FileRequests.create client , “cool”, “/Temp”, deadline

More info at: https://www.dropbox.com/developers/documentation/http/documentation#file_requests-create

get(client, id)

Returns the specified file request.

Example

ElixirDropbox.FileRequests.get client , “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#file_requests-get

list(client)

Returns a list of file requests owned by this user. For apps with the app folder permission, this will only return file requests with destinations in the app folder.

Example

ElixirDropbox.FileRequests.list client

More info at: https://www.dropbox.com/developers/documentation/http/documentation#file_requests-list

update(client, id, title, destination, deadline, open \\ true)

Update a file request.

Example

deadline = %{ “deadline” => “2020-10-12T17:00:00Z” } ElixirDropbox.FileRequests.create client, “”, “cool”, “/Temp”, deadline

More info at: https://www.dropbox.com/developers/documentation/http/documentation#file_requests-list