ElixirDropbox v0.0.7 ElixirDropbox.Files.CopyBatch

Functions

check(client, async_job_id)

Copy multiple files or folders to different locations at once in the user’s Dropbox.

Example

ElixirDropbox.Files.CopyBatch.check(client, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-copy_batch

copy_batch(client, from_path, to_path, allow_shared_folder \\ false, autorename \\ false, allow_ownership_transfer \\ false)

Copy multiple files or folders to different locations at once in the user’s Dropbox.

Example

ElixirDropbox.Files.CopyBatch.copy_batch(client, “/Temp/“, “/Tmp”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-copy_batch