ElixirDropbox v0.0.8 ElixirDropbox.Files.CopyReference

Functions

get(client, path)

Get a copy reference to a file or folder. This reference string can be used to save that file or folder to another user’s Dropbox by passing it to copy_reference/save.

Example

ElixirDropbox.Files.CopyReference.get(client, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-copy_reference-get

save(client, copy_reference, path)

Save a copy reference returned by copy_reference/get to the user’s Dropbox.

Example

ElixirDropbox.Files.CopyReference.get(client, “”, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-copy_reference-save