ElixirDropbox v0.0.8 ElixirDropbox.Files.DeleteBatch

Functions

check(client, async_job_id)

Returns the status of an asynchronous job for delete_batch. If success, it returns list of result for each entry.

Example

ElixirDropbox.Files.DeleteBatch.check(client, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-delete_batch-check

delete_batch(client, path_entries)

Delete multiple files/folders at once.

Example

path_entries = [ %{“path” => “/Temp1”}, %{“path” => “/Temp2” } ] ElixirDropbox.Files.DeleteBatch.delete_batch(client, path_entries)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-delete_batch