ElixirDropbox v0.0.6 ElixirDropbox.Files.ListFolder

Functions

get_latest_cursor(client, path)

get_latest_cursor(Client.t, binary) :: ElixirDropbox.response

A way to quickly get a cursor for the folder’s state.

##Example

ElixirDropbox.Files.ListFolder.get_latest_cursor client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-list_revisions

list_folder(client, path)

list_folder(Client.t, binary) :: ElixirDropbox.response

Starts returning the contents of a folder.

##Example

ElixirDropbox.Files.ListFolder.list_folder client, “/path”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-list_folder

list_folder_continue(client, cursor)

list_folder_continue(Client.t, binary) :: ElixirDropbox.response

Once a cursor has been retrieved from list_folder, use this to paginate through all files and retrieve updates to the folder, following the same rules as documented for list_folder.

##Example

ElixirDropbox.Files.ListFolder.list_folder_continue client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-list_folder-continue

list_revisions(client, path, limit \\ 10)

list_revisions(Client.t, binary, number) :: ElixirDropbox.response

Return revisions of a file.

##Example

ElixirDropbox.Files.ListFolder.list_revisions client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-list_revisions

longpoll(client, cursor)

longpoll(Client.t, binary) :: ElixirDropbox.response

A longpoll endpoint to wait for changes on an account.

##Example

ElixirDropbox.Files.ListFolder.longpoll client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-list_folder-longpoll