ElixirDropbox v0.0.8 ElixirDropbox.Files.UploadSession

Functions

append(client, session_id, close, file, offset \\ 0)

Append more data to an upload session.

Example

ElixirDropbox.Files.UploadSession.append(client, “AAAAAAAADipnX-8-d8-V4g”, false, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-upload_session-append_v2

finish(client, session_id, path, file, offset \\ 0, mode \\ "add", autorename \\ true, mute \\ false)

Finish an upload session and save the uploaded data to the given file path.

Example

ElixirDropbox.Files.UploadSession.finish(client, “AAAAAAAADipnX-8-d8-V4g”, “”, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-upload_session-finish

finish_batch(client, entries)

This route helps you commit many files at once into a user’s Dropbox.

Example

entry = %{ “cursor” => %{ “session_id” => “AAAAAAAADipnX-8-d8-V4g”, “offset” => 0}, “commit” => %{ “path” => “/Homework/math/Matrices.txt” }} entries = [entry] ElixirDropbox.Files.UploadSession.finish_batch client, entries

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-upload_session-finish

finish_batch_check(client, async_job_id)

Returns the status of an asynchronous job for upload_session/finish_batch. If success, it returns list of result for each entry.

Example

ElixirDropbox.Files.UploadSession.finish_batch_check(client, “”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-upload_session-finish_batch-check

start(client, close, file)

Upload sessions allow you to upload a single file in one or more requests, for example where the size of the file is greater than 150 MB.

Example

ElixirDropbox.Files.UploadSession.start(client, false, “/Tmp”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-upload_session-start