ElixirDropbox v0.0.7 ElixirDropbox.Paper.Docs

Functions

docs_archive(client, doc_id)

Marks the given Paper doc as archived.

Example

ElixirDropbox.Paper.Docs.docs_archive client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-create

docs_create(client, import_format, parent_folder_id, file)

Creates a new Paper doc with the provided content.

Example

ElixirDropbox.Paper.Docs.docs_create client, “”, “”, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-create

docs_download(client, doc_id, export_format)

Exports and downloads Paper doc either as HTML or markdown.

Example

ElixirDropbox.Paper.Docs.docs_download client, “”, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-download

docs_list(client, filter_by \\ "docs_created", sort_by \\ "modified", sort_order \\ "descending", limit \\ 100)

Return the list of all Paper docs according to the argument specifications.

Example

ElixirDropbox.Paper.Docs.docs_list client

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-list

docs_update(client, doc_id, doc_update_policy, revision, import_format, file)

Updates an existing Paper doc with the provided content.

Example

ElixirDropbox.Paper.Docs.docs_update client, “”, “”, “”, “”, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-download

get_folder_info(client, doc_id)

Retrieves folder information for the given Paper doc.

Example

ElixirDropbox.Paper.Docs.get_folder_info client, “”, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-get_folder_info

permanently_delete(client, doc_id)

Permanently deletes the given Paper doc.

Example

ElixirDropbox.Paper.Docs.permanently_delete client

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-list