ElixirDropbox v0.0.8 ElixirDropbox.Paper.SharingPolicy

Functions

get(client, doc_id)

Gets the default sharing policy for the given Paper doc.

Example

ElixirDropbox.Paper.SharingPolicy.get client, “”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-sharing_policy-get

set(client, doc_id, sharing_policy)

Sets the default sharing policy for the given Paper doc.

Example

ElixirDropbox.Paper.SharingPolicy.set client, “HAOV9lRfMNj90iGLqMmC7”, sharing_policy More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-sharing_policy-set