ElixirDropbox v0.0.8 ElixirDropbox.Paper.Users

Functions

add(client, doc_id, members, custom_message, quiet)

Allows an owner or editor to add users to a Paper doc or change their permissions using their email address or Dropbox account ID.

Example

members = [%{ “member” => %{ “.tag” => “email”, “email” => “test@test.com” }, “permission_level”: “view_and_comment” } ] ElixirDropbox.Paper.Users.add client, “HAOV9lRfMNj90iGLqMmC7”, members, “”, false

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-users-add

list(client, doc_id, limit \\ 100, filter_by \\ "shared")

Lists all users who visited the Paper doc or users with explicit access.

Example

ElixirDropbox.Paper.Users.list client, “HAOV9lRfMNj90iGLqMmC7”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-users-list

list_continue(client, doc_id, cursor)

Once a cursor has been retrieved from docs/users/list, use this to paginate through all users on the Paper doc.

Example

ElixirDropbox.Paper.Users.list_continue client, “HAOV9lRfMNj90iGLqMmC7”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-users-list-continue

remove(client, doc_id, member)

Allows an owner or editor to remove users from a Paper doc using their email address or Dropbox account ID.

Example

member = %{ “member” => %{ “.tag” => “email”, “email” => “test@test.com” } ElixirDropbox.Paper.Users.remove client, “”, member

More info at: https://www.dropbox.com/developers/documentation/http/documentation#paper-docs-users-remove