ElixirDropbox v0.0.6 ElixirDropbox.Files

Functions

copy(client, from_path, to_path)

Copy a file or folder to a different location in the user’s Dropbox. If the source path is a folder all its contents will be copied.

Example

ElixirDropbox.Files.copy(client, “/Temp/first”, “/Tmp/second”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-copy_v2

copy_batch(client, from_path, to_path)

create_folder(client, path)

create_folder(Client, binary) :: Map

Create folder returns map

Example

ElixirDropbox.Files.create_folder client, “/Path”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-create_folder

create_folder_to_struct(client, path)

create_folder_to_struct(Client, binary) :: Map

Create folder returns Folder struct

Example

ElixirDropbox.Files.create_folder_to_struct client, “/Path”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-create_folder

delete_folder(client, path)

Delete folder returns map

## Example

ElixirDropbox.Files.delete_folder client, “/Path”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-delete

delete_folder_to_struct(client, path)

Delete folder returns Folder struct

Example

ElixirDropbox.Files.delete_folder_to_struct client, “/Path”

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-delete

download(client, path)

get_metadata(client, path, include_media_info \\ false)

get_preview(client, path)

get_thumbnail(client, path, format \\ "jpeg", size \\ "w64h64")

move(client, from_path, to_path)

Move a file or folder to a different location in the user’s Dropbox. If the source path is a folder all its contents will be moved.

Example

ElixirDropbox.Files.move(client, “/Homework/math”, “/Homework/algebra”)

More info at: https://www.dropbox.com/developers/documentation/http/documentation#files-move

restore(client, path, rev)

upload(client, path, file, mode \\ "add", autorename \\ true, mute \\ false)